refactor: split helper into subcomponents
